Yargh, thar be resources in sight\!

+++ Components/src/com/earth2me/components/storage/YamlStore.java
@@ -3,6 +3,7 @@ package com.earth2me.components.storage;
 import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
+import java.util.Map;
 import org.yaml.snakeyaml.DumperOptions;
 import org.yaml.snakeyaml.Yaml;
 import org.yaml.snakeyaml.nodes.Tag;
@@ -15,28 +16,42 @@ import org.yaml.snakeyaml.representer.Representer;
  * @param <T> the type to serialize/store.
  * @author Zenexer
  */
-public class YamlStore<T extends IStorable> implements IBackedStore
+public final class YamlStore<T extends IStorable> implements IBackedStore
 {
 
 	private final transient Yaml yaml;
 	private final transient File file;
 	private final transient Class<?> type;
+	private final transient boolean readonly;
 	private transient T data;
 
 	/**
-	 * Instantiates a new YAML store.  Does NOT load from the file.
+	 * Instantiates a new YAML store. Does NOT load from the file. Defaults to
+	 * readonly.
 	 *
 	 * @param file the file used to back the YAML store.
 	 * @param data an instance to initialize the store.
 	 */
 	public YamlStore(File file, T data)
 	{
+		this(file, data, true);
+	}
+
+	/**
+	 * Instantiates a new YAML store. Does NOT load from the file.
+	 *
+	 * @param file the file used to back the YAML store.
+	 * @param data an instance to initialize the store.
+	 * @param readonly true to prevent saving; otherwise false.
+	 */
+	public YamlStore(File file, T data, boolean readonly)
+	{
 		this.file = file;
 		this.data = data;
+		this.readonly = readonly;
 		this.type = data.getClass();
 
-		// Configure the YAML parser/emitter.
-
+		// Configure the YAML emitter's options.
 		final DumperOptions options = new DumperOptions();
 		options.setPrettyFlow(true);
 		options.setIndent(2);
@@ -48,9 +63,16 @@ public class YamlStore<T extends IStorable> implements IBackedStore
 		options.setDefaultFlowStyle(DumperOptions.FlowStyle.AUTO);
 		options.setDefaultScalarStyle(DumperOptions.ScalarStyle.DOUBLE_QUOTED);
 
+		// Register class tags.
 		final Representer representer = new Representer();
-		representer.addClassTag(type, new Tag("!" + getData().getId()));
+		final Map<String, Class<?>> customTags = data.getCustomClassTags();
+		for (String tag : customTags.keySet())
+		{
+			representer.addClassTag(customTags.get(tag), new Tag("!" + tag));
+		}
+		representer.addClassTag(type, new Tag("!" + data.getId()));
 
+		// Instantiate yaml with all the parameters we just defined.
 		this.yaml = new Yaml(representer, options);
 	}
 
@@ -62,8 +84,14 @@ public class YamlStore<T extends IStorable> implements IBackedStore
 	@Override
 	public boolean save()
 	{
+		if (readonly)
+		{
+			return false;
+		}
+
 		try
 		{
+
 			final FileWriter writer = new FileWriter(file);
 			try
 			{
